:art: Change __init.py__ for app.py
